Transition lambda worker to separate app entrypoint.
This change decouples the lambda worker from the existing kitchensink worker. We introduce a new hook into the harness to register a lambda worker. A separate hook is necessary because the lambdaworker package controls the lifecycle of the worker, where the existing worker factory assumes the harness controls the lifecycle.

This change allows for arbitrary apps to be created with lambda worker invocations instead of long-running harness controlled workers.

Subsequent work to read omes worker flags and provide their values into the lambda factory is deferred to another PR (this would allow lambda workers to be configured like existing WorkerFactory - pass CLI args, they are passed as the input to the factory function. Users can override/set fields as necessary inside their factory function)
